add confirm delete

+++ src/components/pages/Blog.jsx
@@ -22,11 +22,14 @@ const Blog = () => {
     }, []);
 
     const handleDelete = async (slug) => {
-        try {
-            await axios.delete(`/posts/${slug}`);
-            setPosts((prevPosts) => prevPosts.filter((post) => post.slug !== slug)); 
-        } catch (error) {
-            console.error('Error deleting post:', error);
+        const confirmed = window.confirm('Sei sicuro di voler cancellare questo post?');
+        if (confirmed) {
+            try {
+                await axios.delete(`/posts/${slug}`);
+                setPosts((prevPosts) => prevPosts.filter(post => post.slug !== slug));
+            } catch (error) {
+                console.error('Error deleting post:', error);
+            }
         }
     };
     return (
