+++ src/components/pages/Post.jsx
@@ -24,12 +24,15 @@ const Post = () => {
         fetchPost();
     }, [slug]);
 
-    const handleDelete = async () => {
-        try {
-            await axios.delete(`/posts/${slug}`);
-            navigate('/blog'); 
-        } catch (error) {
-            console.error('Error deleting post:', error);
+    const handleDelete = async (slug) => {
+        const confirmed = window.confirm('Sei sicuro di voler cancellare questo post?');
+        if (confirmed) {
+            try {
+                await axios.delete(`/posts/${slug}`);
+                setPosts((prevPosts) => prevPosts.filter(post => post.slug !== slug));
+            } catch (error) {
+                console.error('Error deleting post:', error);
+            }
         }
     };
